Future-proof port-forward, ssh and vnc commands
To support not only the VM resource.
